Hanoi.TowerGame (Towers of Hanoi v0.6.0)

The Hanoi game as a public API & state.

Each individual game should run as one of these. They will exist independently of each other.

Runs as a supervisor and starts a GenServer to process work and an ETS instance to hold state. Usually run as part of an OTP application.

Functions

get_moves(name)

@spec get_moves(name :: atom()) :: list() | :error

Function that for a board with all stones on left pile will move them in a legal way (no big stone on small stone) to the right hand pile

get_number_moves(name)

@spec get_number_moves(name :: atom()) :: non_neg_integer()

Function returning the number of moves that have been played.

get_number_stones(name)

@spec get_number_stones(name :: atom()) :: non_neg_integer()

Function that returns the number of stones on the board.

get_state(name)

@spec get_state(name :: atom()) :: Hanoi.Board.t()

Function returning the current state of the hanoi board

get_state_as_text(name)

@spec get_state_as_text(name :: atom()) :: String.t()

Debug function returning current state of hanoi board in a textual representation.

is_complete(name)

@spec is_complete(name :: atom()) :: boolean() | :error

Funtion returns true if game is complete That is all stones in correct order on right hand pile

move_stone(name, from, to)

@spec move_stone(name :: atom(), from :: atom(), to :: atom()) :: :ok | :error

Funtion that moves a stone between piles. Will error if the stone is bigger than the one at the top of the pile. Increments moves taken.

reset(name, new_stones)

@spec reset(name :: atom(), new_stones :: pos_integer()) :: :ok

start_link(opts)

@spec start_link(opts :: list()) :: {:error, any()} | {:ok, pid()}

Call this function to start up the module. Name and Stones must be provided.